// Frontend/src/modules/chats/domain/valueObjects/chat-enums.ts
export enum ChatChannel {
  WHATSAPP = "WhatsApp"
}

export enum MessageType {
  TEXT = "text",
  IMAGE = "image",
  FILE = "file",
  AUDIO = "audio",
  ALERT = "alert",
  NOTE = "note"
}

export enum ChatFilter {
  CHATS = "chats",
  FILA = "fila",
  CONTATOS = "contatos"
}

export enum ChatSortOption {
  LONGEST_WAIT = "longest_wait",
  SHORTEST_WAIT = "shortest_wait",
  NEWEST = "newest",
  OLDEST = "oldest"
}